Add many changes and run the project successfully
The dataset is split into training set and testing set. Training set is then used to do the training and x train and y train are obtained. The testing set is used to obtain x test and y test, which is in turn used for plotting. Prediction is then made using x test and the values obtained are plotted against y test.